make laxmap hidden
